[Agent] modularize loader helpers
